+++ app/src/main/java/com/elliemoritz/coinbook/data/dao/DebtsOperationsDao.kt
@@ -16,14 +16,14 @@ interface DebtsOperationsDao {
     suspend fun removeAllOperations()
 
     @Query("SELECT * FROM debt_operations WHERE id = :id LIMIT 1")
-    suspend fun getOperation(id: Int): DebtOperationDbModel
+    suspend fun getOperation(id: Long): DebtOperationDbModel
 
     @Query("SELECT * FROM debt_operations WHERE debtId = :debtId LIMIT 1")
-    suspend fun getOperationByDebtId(debtId: Int): DebtOperationDbModel
+    suspend fun getOperationByDebtId(debtId: Long): DebtOperationDbModel
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun addOperation(operation: DebtOperationDbModel)
 
     @Query("DELETE FROM debt_operations WHERE id = :id")
-    suspend fun removeOperation(id: Int)
+    suspend fun removeOperation(id: Long)
 }
