+++ app/src/main/java/com/elliemoritz/coinbook/domain/entities/Limit.kt
@@ -5,7 +5,7 @@ import com.elliemoritz.coinbook.domain.entities.helpers.UNDEFINED_ID
 data class Limit(
     val limitAmount: Int,
     val realAmount: Int,
-    val categoryId: Int,
+    val categoryId: Long,
     val categoryName: String,
-    val id: Int = UNDEFINED_ID
+    val id: Long = UNDEFINED_ID
 )
