+++ app/src/main/java/com/elliemoritz/coinbook/domain/repositories/AlarmsRepository.kt
@@ -5,9 +5,9 @@ import kotlinx.coroutines.flow.Flow
 
 interface AlarmsRepository {
     fun getAlarmsList(): Flow<List<Alarm>>
-    fun getAlarm(id: Int): Flow<Alarm>
+    fun getAlarm(id: Long): Flow<Alarm>
     suspend fun addAlarm(alarm: Alarm)
     suspend fun editAlarm(alarm: Alarm)
     suspend fun removeAlarm(alarm: Alarm)
-    fun getAlarmsCount(): Flow<Int>
+    fun getAlarmsCount(): Flow<Long>
 }
