+++ app/src/main/java/com/elliemoritz/coinbook/domain/repositories/DebtsOperationsRepository.kt
@@ -8,8 +8,8 @@ interface DebtsOperationsRepository {
     fun getOperationsList(): Flow<List<DebtOperation>>
     suspend fun removeAllOperations()
 
-    fun getOperation(id: Int): Flow<DebtOperation>
-    fun getOperationByDebtId(debtId: Int): Flow<DebtOperation>
+    fun getOperation(id: Long): Flow<DebtOperation>
+    fun getOperationByDebtId(debtId: Long): Flow<DebtOperation>
     suspend fun addOperation(operation: DebtOperation)
     suspend fun editOperation(operation: DebtOperation)
     suspend fun removeOperation(operation: DebtOperation)
