+++ app/src/main/java/com/elliemoritz/coinbook/domain/repositories/DebtsRepository.kt
@@ -5,7 +5,7 @@ import kotlinx.coroutines.flow.Flow
 
 interface DebtsRepository {
     fun getDebtsList(): Flow<List<Debt>>
-    fun getDebt(id: Int): Flow<Debt>
+    fun getDebt(id: Long): Flow<Debt>
     suspend fun addDebt(debt: Debt)
     suspend fun editDebt(debt: Debt)
     suspend fun removeDebt(debt: Debt)
